Renaming lib/libcheck to types/check
